+++ src/components/people/TechStackAccordion/index.tsx
@@ -1,7 +1,7 @@
 import { useCallback, useEffect, useRef } from "react";
 
 import { useAccordion } from "@/hooks/people/useAccordion";
-import { useTechStack } from "@/hooks/useTechStack";
+import { useTechStack } from "@/hooks/people/useTechStack";
 
 import chevronIcon from "@/assets/people/chevron.svg";
