+++ src/components/project/ExtendedProjectCard/index.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 
-import { ProjectManagement } from "@/components/project/ProjectManagement/ProjectManagement";
+import { ProjectManagement } from "@/components/project/ProjectManagement";
 
 import * as Styles from "./index.style";
 import { css } from "@emotion/react";
